requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django==1.11
django-model-utils==3.0.0
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
wheel==0.24.0
